HTTP get against httpbin.org seems to work.
We get both headers and body content.  Note that it is meaningful that the
body content is delivered all at once in the buffer.  Consequently it might
be wise not to assume that this is 100% correct in the face of larger bodies
or of transfers that are delivered a part at a time.