HTTP response structures made public, more work.
We have put in prototypes for aio structures, but it really
is time for us to work on the generic input/output support for
aio structures because we need this for the generic http channel
read/write routines.

We could just expose the iov directly, but it looks like a better
approach might be to just nuke the iov, instead passing it via
an extensible nng_aio_set_input()... imagine parameter 0 is the
size of the iov, and then we pass items in pairs (buf & length).