Significant changes.
This changes raw mode so that it is immutable, determined
by the socket constructor.  Additionally, this frees many
protocols from handling the raw mode option, which can now be
handled by the socket core based on flags in the protocol registration
structure.

This also leads to shorter code paths and fewer cases of dealing with
locks, which should ultimately improve performance.

Context support is plumbed all the way through now; we need to add
context support for REP, and other protocols.

Also, we need to document the context API and add some tests.